add InMemoryHistoryManager his own class
